refactor reduce into separate traits
