+++ crates/burn-jit/src/kernel/reduce/mod.rs
@@ -1,23 +1,11 @@
-mod argmax_dim;
-mod argmin_dim;
 mod base;
-mod mean_dim;
-mod naive_reduce_shader;
+mod naive;
 mod prod;
-mod prod_dim;
-mod shared_reduce_shader;
+mod shared;
 mod sum;
-mod sum_dim;
 mod tune;
 
-pub(crate) use argmax_dim::*;
-pub(crate) use argmin_dim::*;
 pub use base::*;
-pub(crate) use mean_dim::*;
-pub use naive_reduce_shader::*;
 pub use prod::*;
-pub(crate) use prod_dim::*;
-pub use shared_reduce_shader::*;
 pub use sum::*;
-pub(crate) use sum_dim::*;
 pub use tune::*;
