+++ crates/burn-jit/src/kernel/reduce/naive/mod.rs
@@ -1,0 +1,7 @@
+pub(crate) mod argmax;
+pub(crate) mod argmin;
+pub(crate) mod base;
+pub(crate) mod mean_dim;
+pub(crate) mod prod_dim;
+pub(crate) mod shader;
+pub(crate) mod sum_dim;
