+++ crates/burn-jit/src/kernel/reduce/shared/base.rs
@@ -1,0 +1,49 @@
+use burn_cube::dialect::{Item, Scope, Variable};
+
+use crate::JitElement;
+
+/// Specifies the reduce dim algorithm in use
+pub trait ReduceDimShared<E: JitElement>: Send + Sync + 'static {
+    /// The reduction accumulator
+    type Accumulator: Copy;
+
+    /// Initialization for shared algorithm
+    fn initialize_shared(
+        scope: &mut Scope,
+        shared_memory_size: u32,
+        write_position: Variable,
+        input_item: Item,
+    ) -> Self::Accumulator;
+
+    /// How to write to shared memory
+    fn write_to_shared(
+        scope: &mut Scope,
+        shared_memory: Self::Accumulator,
+        write_position: Variable,
+        value: Self::Accumulator,
+    );
+
+    /// How to read from input in shared algorithm
+    fn read_from_input(
+        scope: &mut Scope,
+        input: Variable,
+        read_position: Variable,
+        i: Variable,
+    ) -> Self::Accumulator;
+
+    /// How to read from shared memory
+    fn read_from_shared(
+        scope: &mut Scope,
+        shared_memory: Self::Accumulator,
+        read_position: Variable,
+    ) -> Self::Accumulator;
+
+    /// How to assign from shared memory
+    fn assign_shared(
+        scope: &mut Scope,
+        shared_memory: Self::Accumulator,
+        output: Variable,
+        write_position: Variable,
+        shape_reduce_dim: Variable,
+    );
+}
